Generate table.rs using new build.rs
The magic numbers in table.rs are generated in a predictable fashion using a
stream cipher. Rather than having the magic file in source control, generate
the file as needed at build time. The results are always the same, so the
chunking is still as predictable as ever.

cargo test passes