feat: add read support for v2016, v2020 chunkers
In addition to the FastCDC types in the v2016 and v2020 modules, now there
are StreamCDC structs that will read from a boxed Read into a buffer sized
to fit the maximum chunk. While this is convenient for processing large
files, it is a bit slower than using memory-mapped files with a crate such
as memmap2. Added examples that demonstrate using the streaming chunkers.

cargo test passes